requirements.txt
albumentations==1.3.0
einops==0.6.0
future==0.18.2
fvcore==0.1.5.post20221221
h5py==3.1.0
joblib==1.1.0
kornia==0.6.9
loguru==0.6.0
matplotlib==3.5.1
opencv-python==4.7.0.72
Pillow==9.0.1
pytorch-lightning==1.9.0
scikit-image==0.19.1
scikit-learn==1.0.2
tqdm==4.62.3
yacs==0.1.8
torchmetrics==0.11.0
gdown
tensorboardX==2.6
timm==0.6.7
